Introduce Analyze Table and Tuple
